requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
requests
beautifulsoup4
redis
faiss-cpu
numpy
